pyproject.toml
[tool.poetry]
name = "python gui with tkinter"
version = "0.1.0"
description = ""
authors = ["Charchit <charchitregmi10@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pillow = "^10.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
